changes to fix compiler warnings : changed   uint32_t ina219_calValue; to unint16_t

+++ Adafruit_INA219.h
@@ -137,7 +137,8 @@ enum {
   void powerSave(int fd, int on);
 
   uint8_t ina219_i2caddr;
-  uint32_t ina219_calValue;
+  //uint32_t ina219_calValue;
+  uint16_t ina219_calValue;
   // The following multipliers are used to convert raw current and power
   // values to mA and mW, taking into account the current config settings
   uint32_t ina219_currentDivider_mA;
